Modify project structure and update CarFactory to include new tires component
- Model directory removed. CarFactory class now utilizes Car class to create car objects of any model type. Previous related model imports have been removed.
- CarFactory class definitions for each model now include tires in composition and tire_wear array for service checks.
- Type hint for tire_wear array updated from List[int] to List[float].

+++ component/tires/carrigan_tires.py
@@ -2,7 +2,7 @@ from typing import List
 from interface.serviceable import Serviceable
 
 class CarriganTires(Serviceable):
-    def __init__(self, tire_wear: List[int]):
+    def __init__(self, tire_wear: List[float]):
         self.tire_wear = tire_wear
     
     def needs_service(self) -> bool:
